# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + PocketBase backend for QuantumAtlas: paper\ncollection + search + database. Write endpoints and\nthe search/paper read surface require a bearer token\n(PAT or PocketBase session) plus the matching scope.\nSee the auth model docs for the scope vocabulary.",
        "title": "QuantumAtlas API",
        "contact": {
            "name": "QuantumAtlas",
            "url": "https://quantum-atlas.ai"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/db/schema": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Read-only introspection of the QATLAS_POSTGRES_DSN database:\nevery table in schema public (goose_db_version included,\nalphabetical) with row estimate, total size, columns\n(type/nullable/default/is_pk), indexes and constraints.\nAdmin-only: session token + is_admin or is_superadmin\non the users record.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Postgres schema browser",
                "responses": {
                    "200": {
                        "description": "{database, tables:[{name, row_estimate, total_size, columns, indexes, constraints}]}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "postgres registry unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/whoami": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns {login, is_admin, is_user_admin, is_superadmin}\nfor the signed-in session user. login is the github_login,\nfalling back to gitea_login when only that is stamped.\nAll three flags read the users-record role flags (the\nauth.admin_logins / superadmin_logins YAML lists only\nseed them at boot): is_admin and is_user_admin are true\nwhen the record holds is_admin or is_superadmin; the\nSPA uses them to decide which admin surfaces to render.\nNon-admins get false flags rather than a 403.\nSession-token auth only (PAT auth\nrefused, same as /api/pat).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin whoami",
                "responses": {
                    "200": {
                        "description": "{login, is_admin, is_user_admin, is_superadmin}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "PAT auth not accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003ctoken\u003e\" — two credential shapes\naccepted: a Personal Access Token\n(`Authorization: Bearer qat_...`, minted at\n`/pat` after GitHub OAuth login), or the\nconfig-loaded system PAT (set\n`system_pat.token` in config.yaml on the\nserver, send the plaintext as\n`Authorization: Bearer \u003cvalue\u003e`).\nBrowser callers are authenticated through\npb.authStore (no copy step) — only non-browser\ncallers need an explicit bearer.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}